requirements.txt
celery==5.2.7
Django==4.1.2
django-environ==0.9.0
djangorestframework==3.14.0
flower==1.2.0
Scrapy==2.7.0
scrapy-djangoitem==1.1.1
